refactor(suite-native): tighten wrap analytics types, deps and docs
